Update simple_processor.c
Found a nasty bug for sr00 format - If the size is 256 (0 1), it was computing the length as 0 instead of 1 - looked like it had a problem getting out of char operations into int operations.  So I clarified the fix, and volla! it works.

Only would happen rarely, causing the listener to freeze up on files, and probably crash
Fixed in processor.c in the main code as well.